update : fork 시 aux 도 duplicate 하기 위해 uninit_duplicate_aux() 함수 생성 1. setup_stack 시 런타임 스택이라는 것을 알 수 있도록 유형에 marker_0 추가 2. check_addr 함수에서 vm에서는 pml4에서 페이지를 찾지 못해도 exit하지 않고 page_fault하도록 변경 3. uninit_destroy 함수에서 가지고 있는 aux free 해주도록 추가

+++ include/userprog/process.h
@@ -15,7 +15,7 @@ struct load_info {
     struct file *file;
     size_t page_read_bytes;
     size_t page_zero_bytes;
-    int32_t ofs;
+    off_t ofs;
 };
